moved settings widgets
